+++ user-service/src/main/java/br/com/avf/user/protocol/UserResponse.java
@@ -3,10 +3,12 @@ package br.com.avf.user.protocol;
 import lombok.Data;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Data
 @ToString
 public class UserResponse {
-    private Integer id;
+    private Long id;
     private String name;
-    private Integer balance;
+    private BigDecimal balance;
 }
