+++ user-service/src/main/java/br/com/avf/user/protocol/UserTransactionResponse.java
@@ -3,10 +3,12 @@ package br.com.avf.user.protocol;
 import lombok.Data;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Data
 @ToString
 public class UserTransactionResponse {
-    private Integer userId;
-    private Integer amount;
+    private Long userId;
+    private BigDecimal amount;
     private UserTransactionStatus status;
 }
